move traces_test.go tests to tracesgen_test.go
